Refactor.

+++ src/app/build.rs
@@ -13,6 +13,57 @@ const INFO_TEXT_SIZE: f32 = 18.0;
 pub(super) mod editor;
 pub(super) mod file_list;
 
+#[derive(Debug, Default, Clone)]
+pub struct NewFileState {
+    filename: String,
+    new_password: String,
+    confirm_password: String,
+    font_size: f32,
+    data_dir: String,
+}
+
+impl NewFileState {
+    pub fn new(filename: String, font_size: f32, data_dir: String) -> Self {
+        Self {
+            filename,
+            new_password: "".to_string(),
+            confirm_password: "".to_string(),
+            font_size,
+            data_dir,
+        }
+    }
+    pub fn filename(&self) -> &String {
+        &self.filename
+    }
+}
+
+#[derive(Debug, Default, Clone)]
+pub struct EncryptedFileState {
+    filename: String,
+    ciphertext: String,
+    password: String,
+    new_password: String,
+    confirm_password: String,
+    font_size: f32,
+    data_dir: String,
+}
+
+impl EncryptedFileState {
+    pub fn new(filename: String, ciphertext: String, font_size: f32, data_dir: String) -> Self {
+        Self {
+            filename,
+            ciphertext,
+            font_size,
+            data_dir,
+            ..Default::default()
+        }
+    }
+
+    pub fn filename(&self) -> &String {
+        &self.filename
+    }
+}
+
 impl MyApp {
     pub(super) fn main_layout(
         &mut self,
@@ -24,17 +75,18 @@ impl MyApp {
         ));
         ui.with_layout(egui::Layout::left_to_right(egui::Align::TOP), |ui| {
             self.build_file_list(FILE_LIST_WIDTH, ctx, ui);
-            match self.content.clone() {
-                Content::NewFile(filename) => {
+            match &mut self.content {
+                Content::NewFile(ref mut new_file_state) => {
                     ui.with_layout(egui::Layout::top_down(egui::Align::Center), |ui| {
                         ui.allocate_space(Vec2::new(0.0, PASSWORD_SCREEN_TOP_SPACE));
-                        self.build_uninitialized_file(filename, ctx, ui);
+                        self.next_content = Self::build_uninitialized_file(new_file_state, ctx, ui);
                     });
                 }
-                Content::Encrypted(ref filename, ref iv, ref data, ref mac) => {
+                Content::Encrypted(ref mut encrypted_file_state) => {
                     ui.with_layout(egui::Layout::top_down(egui::Align::Center), |ui| {
                         ui.allocate_space(Vec2::new(0.0, PASSWORD_SCREEN_TOP_SPACE));
-                        self.build_encrypted_file(filename, ctx, ui, iv, data, mac);
+                        self.next_content =
+                            Self::build_encrypted_file(encrypted_file_state, ctx, ui);
                     });
                 }
                 Content::None => {
@@ -48,7 +100,7 @@ impl MyApp {
                     );
                 }
                 Content::PlainText(editor_state) => {
-                    self.build_editor(&editor_state, ctx, ui);
+                    Self::build_editor(&mut self.next_content, editor_state, ctx, ui);
                 }
                 Content::Error(err) => {
                     ui.with_layout(
@@ -77,22 +129,20 @@ impl MyApp {
     fn clear_editor_input_fields(&mut self) {
         self.password = "".to_string();
         self.new_password = ("".to_string(), "".to_string());
-        self.edited_text = "".to_string();
     }
 
     pub(super) fn build_uninitialized_file(
-        &mut self,
-        filename: String,
+        new_file_state: &mut NewFileState,
         ctx: &egui::Context,
         ui: &mut egui::Ui,
-    ) {
+    ) -> Option<Content> {
         ui.add(
-            TextEdit::singleline(&mut self.new_password.0)
+            TextEdit::singleline(&mut new_file_state.new_password)
                 .password(true)
                 .hint_text("New Password"),
         );
         ui.add(
-            TextEdit::singleline(&mut self.new_password.1)
+            TextEdit::singleline(&mut new_file_state.confirm_password)
                 .password(true)
                 .hint_text("Confirm Password"),
         );
@@ -100,7 +150,7 @@ impl MyApp {
         if ui
             .button(egui::WidgetText::RichText(
                 RichText::from("Create").size(18.0).color(
-                    if self.new_password.0 == self.new_password.1 {
+                    if new_file_state.new_password == new_file_state.confirm_password {
                         Color32::BLACK
                     } else {
                         Color32::WHITE.gamma_multiply(0.3)
@@ -110,26 +160,27 @@ impl MyApp {
             .clicked()
             || ctx.input(|i| i.key_pressed(egui::Key::Enter))
         {
-            if self.new_password.0.len() > 0 && self.new_password.0 == self.new_password.1 {
-                self.content = Content::PlainText(EditorState::empty(filename));
-                self.show_passage_operation_buttons = false;
-                self.password = self.new_password.0.clone();
-                self.new_password = ("".to_string(), "".to_string());
+            if new_file_state.new_password.len() > 0
+                && new_file_state.new_password == new_file_state.confirm_password
+            {
+                return Some(Content::PlainText(EditorState::empty(
+                    new_file_state.filename.clone(),
+                    new_file_state.font_size,
+                    new_file_state.new_password.clone(),
+                    new_file_state.data_dir.clone(),
+                )));
             }
         }
+        return None;
     }
 
     pub(super) fn build_encrypted_file(
-        &mut self,
-        filename: &String,
+        encrypted_file_state: &mut EncryptedFileState,
         ctx: &egui::Context,
         ui: &mut egui::Ui,
-        iv: &str,
-        data: &str,
-        mac: &str,
-    ) {
+    ) -> Option<Content> {
         ui.add(
-            TextEdit::singleline(&mut self.password)
+            TextEdit::singleline(&mut encrypted_file_state.password)
                 .password(true)
                 .hint_text("Password"),
         );
@@ -141,39 +192,40 @@ impl MyApp {
             .clicked()
             || ctx.input(|i| i.key_pressed(egui::Key::Enter))
         {
-            match PlainText::decrypt(&self.password, iv, data, mac) {
+            match PlainText::decrypt(
+                &encrypted_file_state.password,
+                &encrypted_file_state.ciphertext,
+            ) {
                 Ok(plaintext) => {
-                    if let Some(text) = plaintext.get_first_passage_text() {
-                        self.edited_text = text;
-                    }
-                    self.content =
-                        Content::PlainText(EditorState::new(filename.clone(), plaintext));
-                    self.dirty = false;
-                    self.add_new_passage = None;
-                    self.editing_passage_name = None;
-                    self.confirm_delete_passage = None;
-                    self.show_passage_operation_buttons = false;
+                    let editor_state = EditorState::new(
+                        encrypted_file_state.filename.clone(),
+                        plaintext,
+                        encrypted_file_state.password.clone(),
+                        encrypted_file_state.font_size,
+                        encrypted_file_state.data_dir.clone(),
+                    );
+                    return Some(Content::PlainText(editor_state));
                 }
                 Err(err) => {
-                    self.content = Content::Error(format!("{:?}", err));
+                    return Some(Content::Error(format!("{:?}", err)));
                 }
             }
         }
         ui.allocate_space(Vec2::new(0.0, 100.0));
         ui.add(
-            TextEdit::singleline(&mut self.new_password.0)
+            TextEdit::singleline(&mut encrypted_file_state.new_password)
                 .password(true)
                 .hint_text("New Password"),
         );
         ui.add(
-            TextEdit::singleline(&mut self.new_password.1)
+            TextEdit::singleline(&mut encrypted_file_state.confirm_password)
                 .password(true)
                 .hint_text("Confirm Password"),
         );
         if ui
             .button(
                 egui::WidgetText::RichText(RichText::from("Change Password").size(18.0)).color(
-                    if self.new_password.0 == self.new_password.1 {
+                    if encrypted_file_state.new_password == encrypted_file_state.confirm_password {
                         Color32::BLACK
                     } else {
                         Color32::WHITE.gamma_multiply(0.3)
@@ -182,22 +234,26 @@ impl MyApp {
             )
             .clicked()
         {
-            if self.new_password.0 == self.new_password.1 {
-                match PlainText::decrypt(&self.password, iv, data, mac) {
+            if encrypted_file_state.new_password == encrypted_file_state.confirm_password {
+                match PlainText::decrypt(
+                    &encrypted_file_state.password,
+                    &encrypted_file_state.ciphertext,
+                ) {
                     Ok(plaintext) => {
-                        let ciphertext = plaintext.encrypt(&self.new_password.0);
-                        let path =
-                            PathBuf::from(self.data_dir.clone()).join(format!("{}.safe", filename));
+                        let ciphertext = plaintext.encrypt(&encrypted_file_state.new_password);
+                        let path = PathBuf::from(encrypted_file_state.data_dir.clone())
+                            .join(format!("{}.safe", encrypted_file_state.filename));
                         std::fs::write(path, &ciphertext).unwrap();
-                        self.content =
-                            Content::Success("Password changed successfully".to_string());
-                        self.clear_editor_input_fields();
+                        return Some(Content::Success(
+                            "Password changed successfully".to_string(),
+                        ));
                     }
                     Err(err) => {
-                        self.content = Content::Error(format!("{:?}", err));
+                        return Some(Content::Error(format!("{:?}", err)));
                     }
                 }
             }
         }
+        return None;
     }
 }
